Splitting security question enroll tests
